feat: lora.* package layout, daemon supervisor, B210 stock-UHD compatibility, qa fixes
Bump gnuradio4 submodule (see submodule's squashed commit).

Package layout — bridges, hwtests, cal:
- lora.bridges.meshcore.{state,repeater,companion} + meshcore-bridge
  CLI entry + lora.identity wiring; migrate-meshcore-json one-shot tool.
- lora.hwtests.{matrix,report,harness} + decode/scan/tx/bridge mode
  runners + standalone diagnostics + 'lora hwtest' CLI tree (Phase 5A-5E).
- lora.cal package + 'lora cal pwr|tinysa|aclr-sa' CLI + smoke tests
  (Phase 6); later extracted to gretel/gr4-rf-tools.

Repo cleanup:
- Deprecation banner on legacy lora_test/bridge_live/perf/cal scripts.
- Delete scripts/apps + scripts/lib + scripts/tests; drop dead code
  (soak_compare, lora decode CLI); drop meshtastic stub.
- docs: scripts README install/run/migration; schema multi-source +
  validation contract; apps/config.toml port-mapping header; tx_gain
  35.5 + preamble_len cleanup.

Daemon supervisor + observability (Phase 8):
- SupervisorConfig parser for [startup].services + KNOWN_SERVICES
  whitelist + service_argv mapper.
- Supervisor class with spawn + signal-driven drain;
  SIGKILL-after-grace; fail-fast for child rc=0/N.
- 'lora daemon' CLI + commented [startup] example;
  --config only forwarded to subcommands that accept it.
- Replay upstream config CBOR to new subscribers; observable startup
  logging + namespace standardization.
- Overlay lora_frame.cr from carrier so dispatch stops dropping frames.
- Orphan-process cleanup hardening across daemon + hwtests; assert_alive
  subprocess watcher; structured terminate handler captures UHD/libusb
  crash text.
- Migrate spawn_lora_agg shim to lora-core daemon; remove parallel-run
  port shim (default :5555).
- Revert FPGA workaround + SoapySDR enumerate probe.
- Bridges: log meshcore-cli stderr/exit on empty stdout.
- core/config: treat existing file as creatable when parent dir writable.

QA — SoapyBlock hardening + B210 stock-UHD compatibility + build cleanup:
- apps/: master_clock_rate 24 MHz across TX dual-sink, RX dual-source,
  scan default. 32 MHz exceeded stock-UHD 2T2R ceiling of 30.72 MHz
  for B210 (uhd-oc fork relaxed via AD9361 overclock; stock enforces).
- apps/: single-BW [62500] for MeshCore EU decode; multi-BW caused
  cross-BW false-locks.
- apps/: drop duplicate apply_startup_config + CMD_*_DEFAULT_FLOOD_SCOPE
  handlers; pin SET-FLOOD-SCOPE write-through.
- blocks/lora ChannelActivityDetector::detectMultiSf — [[maybe_unused]]
  on require_both; document AND-then-OR-fallback contract.
- blocks/lora ScanController::probeStep — drop vestigial isHot/isRefine.
- blocks/lora Splitter, SpectrumTap — explicit static_cast<size_t> on
  tag relIndex for multi-tag publishTag.
- test/qa_lora_scan_classification — drop rAnd assertion.
- .envrc untracked (machine-local direnv); .gitignore updated.

+++ blocks/include/gnuradio-4.0/lora/SpectrumTap.hpp
@@ -68,7 +68,7 @@ struct SpectrumTap : gr::Block<SpectrumTap, gr::NoTagPropagation> {
         // the CircularBuffer double-reserve abort (the OutputSpan ctor
         // already reserved the whole tag buffer).
         for (const auto& [relIndex, tagMapRef] : inSpan.tags()) {
-            outSpan.publishTag(tagMapRef.get(), relIndex);
+            outSpan.publishTag(tagMapRef.get(), static_cast<std::size_t>(relIndex));
         }
 
         // passthrough: copy input to output
